+++ web/vueapp/src/pages/team/single.vue
@@ -159,11 +159,14 @@
                         </v-download-excel>
                     </el-col>
                     <el-col class="team-action-buttons">
-                        <v-link-button v-if="!team.paid" plain :route="{ name: 'etc.admin', query: { type: 'payment', checked: [this.team.id] } }">
+                        <v-link-button icon="el-icon-bell" v-if="!team.paid" plain :route="{ name: 'etc.admin', query: { type: 'payment', checked: [this.team.id] } }">
                             {{ $t('etc.admin.send_payment_reminder') }}
                         </v-link-button>
-                        <v-link-button v-if="!team.hasPlayers"  plain :route="{ name: 'etc.admin', query: { type: 'player_list', checked: [this.team.id] } }">
+                        <v-link-button icon="el-icon-tickets" v-if="!team.hasPlayers"  plain :route="{ name: 'etc.admin', query: { type: 'player_list', checked: [this.team.id] } }">
                             {{ $t('etc.admin.send_player_list_reminder') }}
+                        </v-link-button>
+                        <v-link-button icon="el-icon-message" plain :route="{ name: 'etc.admin', query: { type: 'email', checked: [this.team.id] } }">
+                            {{ $t('etc.admin.send_email_reminder') }}
                         </v-link-button>
                     </el-col>
                 </el-row>